+++ examples/cmp/cpp/gen.cpp
@@ -1,22 +1,24 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-int main(int argc,char* argv[]) {
+int main(int argc, char* argv[]) {
     // Generator for Maximum Subarray Problem
 
-    // quicktest passes a seed as an argument
+    // quicktest passes a seed as an argument: $ ./main seed testcase
     int seed = stoi(string(argv[1]));
     srand(seed);
 
-    int n = rand() % int(1e3) + 1;
+    const int N = int(1e3);
+    const int Ai = int(1e2);
+
+    int n = rand() % N + 1;
     cout << n << "\n";
 
     default_random_engine generator;
-    uniform_int_distribution<int> distribution(-int(1e2), int(1e2));
+    uniform_int_distribution<int> distribution(-Ai, Ai);
 
     for(int i = 0; i < n; i++) {
-        if(i > 0)
-            cout << " ";
+        if(i > 0) cout << " ";
         cout << distribution(generator);
     }