+++ src/runner/cmd.rs
@@ -17,6 +17,7 @@ use super::types::{CPStatus, StatusResponse};
 
 pub fn execute_program(
     timeout: u32,
+    memory_limit: u64,
     testcase: u32,
     commands: Vec<&str>,
     stdin: Option<PathBuf>,
@@ -54,9 +55,9 @@ pub fn execute_program(
     let mut res_status = CPStatus::AC;
 
     if let Ok(child_output) = child {
-        // TODO: add memory_limit method
         let response = child_output
             .controlled_with_output()
+            .memory_limit(memory_limit as usize) // bytes
             .time_limit(Duration::from_millis(timeout as u64))
             .terminate_for_timeout()
             .wait();
@@ -77,7 +78,19 @@ pub fn execute_program(
 
                     res_status = CPStatus::AC;
                 } else {
-                    res_status = CPStatus::RTE;
+                    #[cfg(unix)]
+                    if let Some(6) = output.status.signal() {
+                        res_status = CPStatus::MLE;
+                    } else {
+                        res_status = CPStatus::RTE;
+                    }
+
+                    #[cfg(windows)]
+                    if let Some(3) = output.status.code() {
+                        res_status = CPStatus::MLE;
+                    } else {
+                        es_status = CPStatus::RTE;
+                    }
                 }
             } else {
                 res_status = CPStatus::TLE;