+++ src/subcommand/cmd_example.rs
@@ -8,7 +8,10 @@ use exitfailure::ExitFailure;
 
 use crate::{
     cli::structures::ExampleCommand,
-    views::example::{show_examples_check_cmd, show_examples_cmp_cmd, show_examples_stress_cmd},
+    views::example::{
+        show_examples_check_cmd, show_examples_cmp_cmd, show_examples_output_cmd,
+        show_examples_setup_config_cmd, show_examples_stress_cmd,
+    },
 };
 
 pub fn run(command: &ExampleCommand) -> Result<(), ExitFailure> {
@@ -18,6 +21,10 @@ pub fn run(command: &ExampleCommand) -> Result<(), ExitFailure> {
         show_examples_stress_cmd();
     } else if command.check {
         show_examples_check_cmd();
+    } else if command.output {
+        show_examples_output_cmd();
+    } else if command.setup {
+        show_examples_setup_config_cmd();
     }
     Ok(())
 }